+++ packages/git-service-koa/src/index.ts
@@ -1,76 +1,42 @@
-import { createController, IGenericDriverOptions, LogicController } from "git-service";
+import { createController, IGenericDriverOptions, IRequestData, LogicController } from "git-service";
 import { Middleware } from "koa";
 
-export * from "git-service";
 export default createKoaMiddleware;
 
 /**
- * Creates a handler for direct use with a server instance.
- *
- * @param controller LogicController instance used to serve request.
- * @param keyName Where to store data in `Context.state`.
- *                Defaults to `"gitService"`.
- */
-export function createKoaMiddleware(
-  controller: LogicController,
-  keyName?: string | symbol,
-): Middleware;
-/**
- * Creates a handler for direct use with a server instance.
- *
- * @param optionsOrOrigin Origin or options for a new controller instance.
- * @param keyName Where to store data in `Context.state`.
- *                Defaults to `"gitService"`.
- */
-export function createKoaMiddleware(
-  optionsOrOrigin: string | IGenericDriverOptions,
-  keyName?: string | symbol,
-): Middleware;
-/**
- * Creates a handler for direct use with a server instance.
+ * Creates a handler for use with a koa instance.
  *
  * @param controller LogicController instance used to serve request.
  * @param setup Setup controller before use.
- * @param keyName Where to store data in `Context.state`.
- *                Defaults to `"gitService"`.
  */
 export function createKoaMiddleware(
   controller: LogicController,
   setup?: (controller: LogicController) => any,
-  keyName?: string | symbol,
 ): Middleware;
 /**
- * Creates a handler for direct use with a server instance.
+ * Creates a handler for use with a koa instance.
  *
  * @param optionsOrOrigin Origin or options for a new controller instance.
  * @param setup Setup controller before use.
- * @param keyName Where to store data in `Context.state`.
- *                Defaults to `"gitService"`.
  */
 export function createKoaMiddleware(
   optionsOrOrigin: string | IGenericDriverOptions,
   setup?: (controller: LogicController) => any,
-  keyName?: string | symbol,
 ): Middleware;
 /**
- * Creates a handler for direct use with a server instance.
+ * Creates a handler for use with a koa instance.
  *
- * @param controllerOrOptions LogicController instance used to serve request, or
- *                            options to use with a new controller.
- * @param setup Setup controller before use. Usefull if options is supplied
- *              instead of an existing controller.
- * @param keyName Where to store data in `Context.state`.
- *                Defaults to `"gitService"`.
+ * @param controllerOrOptions LogicController instance or options for a new
+ *                            instance.
+ * @param setup Setup controller before use.
  */
 export function createKoaMiddleware(
-  controllerOrOptions: LogicController | IGenericDriverOptions | string,
-  setup?: ((controller: LogicController) => any) | string | symbol,
-  keyName?: string | symbol,
+  controllerOrOptions: string | IGenericDriverOptions | LogicController,
+  setup?: (controller: LogicController) => any,
 ): Middleware;
 export function createKoaMiddleware(
   controllerOrOptions: LogicController | IGenericDriverOptions | string,
-  setup?: ((controller: LogicController) => any) | string | symbol,
-  keyName: string | symbol = "gitService",
+  setup?: (controller: LogicController) => any,
 ): Middleware {
   let controller: LogicController;
   if (controllerOrOptions instanceof LogicController) {
@@ -80,18 +46,16 @@ export function createKoaMiddleware(
     controller = createController(controllerOrOptions);
   }
   else {
-    throw new TypeError("argument `controller` must be of type 'object'.");
+    throw new TypeError("argument `controllerOrOptions` must be of type 'object' or 'string'.");
   }
   if (typeof setup === "function") {
     setup(controller);
   }
-  else if (typeof setup === "string" || typeof setup === "symbol") {
-    keyName = setup;
-    setup = undefined;
-  }
+  const notAllowed = new Set(["content-length", "content-type"]);
   return async(context, next) => {
+    let request: IRequestData;
     try {
-      const request = await controller.create(
+      request = await controller.create(
         context.req,
         context.req.headers,
         context.method,
@@ -99,15 +63,20 @@ export function createKoaMiddleware(
       );
       request.state = context.state;
       request.state.koa = context;
-      request.state[keyName] = request;
-      const response = await controller.serve(request);
-      response.headers.forEach((h, v) => context.set(h, v));
-      context.status = response.statusCode;
-      context.message = response.statusMessage;
-      context.body = response.body;
+      await controller.serve(request);
     } catch (error) {
       return context.throw(error);
     }
-    return next();
+    const response = request.response;
+    if (response.statusCode === 404) {
+      return next();
+    }
+    if (response.statusCode >= 400) {
+      response.headers.forEach((h, v) => !notAllowed.has(h.toLowerCase()) && context.set(h, v));
+      return context.throw(response.statusCode);
+    }
+    response.headers.forEach((h, v) => context.set(h, v));
+    context.status = response.statusCode;
+    context.body = response.body;
   };
 }